00study/web/demo_flask/requirements.txt
Faker==15.3.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
Flask==2.2.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
Pygments==2.13.0
SQLAlchemy==1.4.44
WTForms==3.0.1
Werkzeug==2.2.2
alembic==1.8.1
bleach==5.0.1
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
dominate==2.7.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
itsdangerous==2.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
visitor==0.1.3
webencodings==0.5.1
yarl==1.8.1
zipp==3.10.0
